blogs/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
tortoise-orm>=0.17.7
asyncpg>=0.24.0
python-multipart>=0.0.5
python-dotenv>=0.19.0